Decoupling DataItem from Recipe
